Move CRD api directory into pkg directory
