+++ PROJECT
@@ -12,7 +12,7 @@ resources:
     namespaced: true
   domain: greymatter.io
   kind: Mesh
-  path: github.com/greymatter-io/operator/api/v1alpha1
+  path: github.com/greymatter-io/operator/pkg/api/v1alpha1
   version: v1alpha1
   webhooks:
     defaulting: true
